Allow handler-function to return a Promise of any
When the handler function returns a Promise of non-void,
the function registering the handler would have to
create a wrapper-lambda-function to await and void the
result of the handler-Promise.
This should not be necessary as for this module it's
not relevant what is actually returned by the Promise
but just whether the Promise has been resolved or rejected.
Therefore the return type of the handler function is
changed to void | Promise<any>.